requirements.txt
python-binance==1.0.16
pandas==1.5.3
scipy==1.10.1